% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gaQSAR_dcv.R
\name{print.gaQSAR_dcv}
\alias{print.gaQSAR_dcv}
\title{Print method for gaQSAR_dcv objects}
\usage{
\method{print}{gaQSAR_dcv}(x, ...)
}
\arguments{
\item{x}{An object of class "gaQSAR_dcv" (returned from \code{gaDoubleCrossValidation()}).}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the object.
}
\description{
Display a brief overview of nested cross-validation results.
}
\seealso{
\code{\link[=summary.gaQSAR_dcv]{summary.gaQSAR_dcv()}}, \code{\link[=gaDoubleCrossValidation]{gaDoubleCrossValidation()}}
}
