% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaPermutationTest.R
\name{print.gaQSAR_permTest}
\alias{print.gaQSAR_permTest}
\title{Print method for gaQSAR_permTest objects}
\usage{
\method{print}{gaQSAR_permTest}(x, ...)
}
\arguments{
\item{x}{A \code{gaQSAR_permTest} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for gaQSAR_permTest objects
}
