% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaPermutationTest.R
\name{summary.gaQSAR_permTest}
\alias{summary.gaQSAR_permTest}
\title{Summary method for gaQSAR_permTest objects}
\usage{
\method{summary}{gaQSAR_permTest}(object, ...)
}
\arguments{
\item{object}{A \code{gaQSAR_permTest} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the input object.
}
\description{
Summary method for gaQSAR_permTest objects
}
